+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Unreleased
 
-## Added
-
 ## Fixed
 
-## Changed
+- Bugfix release
 
 # 0.6.160 (2021-12-03 / 1857689)